More discovery, cleaner keys, more flattening
Separate UserParameter keys for fetching data and doing discovery. This allows
the keys specified in the template to look a little cleaner.

Renamed the --lld-zones to --lld-http-zones and added support for discovering
slab zones, stream zones, as well as upstreams for http and stream. Upstreams
are discovered with two macros, _NAME (of the upstream) and _PEER (each server
in the upstream) to allow for fetching of metrics for each upstream server.

The flatten() function was replaced with a flatten_json() function, that will
flatten absolutely everything in the json structure, and will for the arrays
(currently only peers: [] from what I can see from my NGINX Plus status output)
use the value in 'name' if it exists, otherwise the array index, in the
flattened result.